move to a separate ffi file
