+++ src/lib.rs
@@ -1,2 +1,4 @@
-pub mod bandersnatch_vrfs;
-pub use bandersnatch_vrfs::*;
+mod bandersnatch_vrfs;
+
+pub mod ffi;
+pub use ffi::*;
